web/requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
mysql-connector-python==9.1.0
Jinja2==3.1.4
Werkzeug==3.1.3
